feat: Add `ALLOW_ANONYMOUS_WRITE` to settings
This commit adds a new `ALLOW_ANONYMOUS_WRITE` setting that can be set via
envvar. This is particularly handy when interacting with emeis without
an OIDC provider (during setup or dev). As well as some cornercases,
where you actually want to anonymous users to be able to write certain data.