feat(auth): introduce user factory setting and oidc user object
In OIDC RP projects such as this, we should not use an user model directly
to represent a user in a request. Instead, we use a proxy object (OIDCUser)
to represent the entity doing the request.

This is mainly done for compatibility with other projects that do it the same
way. They may or may not share the same user model.

To facilitate compatibility even further, we introduce a setting named
EMEIS_OIDC_USER_FACTORY, which can be overwritten via environment variables
if the need arises.