feat(search): enable forcing models to be monolingual in search
When users can manage models, they normally only enter data in a single
language.  When people with other language settings then search the same
data, the lookup will happen in the searching user's locale, which will
not return the desired results.

For such cases, we allow forcing a model to a given, fixed,
language. Then, the search on these models will always use the same
language.

This corresponds to a similar change in ember-emeis (our frontend), where
the setting forces the code to store data in the fixed language. Together,
this can then be used to make a multilingual model behave just as it
were monolingual.