feat(scopes): denormalize full_name and use it for sorting
The full_name attribute on the Scopes was dynamically generated on access.
This is rather inefficient, as it may cause a bunch of DB hits, especially
in deep hierarchies.

Additionally, most of the time we want scopes to be sorted in a
meaningful manner (namely, by hierarchical name). MPTTModel supports
this, but as the ordering may differ between languages, we cannot use
that functionality. This is a secondary use of the denormalized full_name.

This also required a fix in the case insensitive ordering code, as
localized fields were not treated correctly yet (casting directly to
text instead of extracting the correct language and using that to sort)