refactor: drop django-mptt in favour of django-tree-queries
Django-Tree-Queries works with recursive CTEs (common table expressions) to deal
with hierarchical relationships, generally achieving the same result, but without
the `right`, `left` and `tree_id` fields.

Django-MPTT has been unmaintained for several years, and we're suspecting some
subtle bugs in it that we can't really reproduce or verify right now. Anyway it's
a good time to get rid of it.

Note: Some old migrations needed to be changed, as we cannot refer to
MPTT fields anymore. They were essentially regular foreign keys however,
so this won't break the database.

BREAKING! This breaks if Emeis is used as a Django app directly, and depends on
any of the MPTT API / features.