feat(models): (re)introduce some convenience methods for working with scope trees
The Django-MPTT module provided some useful methods that are not available anymore
with django-tree-queries. Luckily, it's relatively easy to provide workarounds.

Note that they might not have the same performance/efficiency as the
MPTT variants, and could possibly be built in a better way. However,
let's keep it to the motto "first make it right, then fast, then
pretty"