refactor: drop django-tree-queries & implement array-based tree functions
The Django-Tree-Queries, together with some visibility queries, causes
a ton of heavy performance problems. Therefore, we now implement the
tree functionality for Scopes by using an array field to "cache" all
the parent PKs of every scope.

This allows us to write some much more efficient queries that don't
rely on CTEs, and enables subqueries where we previously had to explode
a QS to perform the same functionality.