chore: drop related name on base model `created_by_user`
This caused, for example, on the `User` model to set an `acl`
attribute, pointing to the creator.

This is very confusing, for example if searching for all users within a scope:

>>> User.objects.filter(acl__scope__in=scope_list)

This would return the user who has *created* all those ACLs, instead of the
users who are in fact *granted* the ACL.

The correct way would have been

>>> User.objects.filter(acls__scope__in=scope_list)

So to avoid this confusion (and others), we'll remove the related name.